deprecated `apply` and (current) `handler`. Renamed `asHandler` to `handler`.
